+++ gems/aws-sdk-core/lib/aws-sdk-core/rpc_v2/cbor_engine.rb
@@ -1,18 +1,17 @@
 # frozen_string_literal: true
 
-require_relative 'encoder'
-require_relative 'decoder'
+require_relative '../cbor'
 
 module Aws
-  module Cbor
+  module RpcV2
     # Pure Ruby implementation of CBOR encode and decode
     module CborEngine
       def self.encode(data)
-        Encoder.new.add(data).bytes
+        Cbor::Encoder.new.add(data).bytes
       end
 
       def self.decode(bytes)
-        Decoder.new(bytes.force_encoding(Encoding::BINARY)).decode
+        Cbor::Decoder.new(bytes.force_encoding(Encoding::BINARY)).decode
       end
     end
   end